+++ src/terminal.rs
@@ -112,6 +112,8 @@ pub(crate) enum TerminalCommand<'a> {
     ScrollContentUp(usize),
 
     DecsetShowCursor,
+    DecsetMouseTracking,
+    DecrstMouseTracking,
     DecsetFocusReporting,
     DecrstFocusReporting,
     DecsetBracketedPaste,
@@ -184,6 +186,8 @@ pub(crate) trait Output {
             QueryCursorPosition => write(self, b"\x1b[6n"),
             ScrollContentUp(lines) => scroll_content_up(self, lines),
             DecsetShowCursor => write(self, b"\x1b[?25h"),
+            DecsetMouseTracking => write(self, b"\x1b[?1000h"),
+            DecrstMouseTracking => write(self, b"\x1b[?1000l"),
             DecsetFocusReporting => write(self, b"\x1b[?1004h"),
             DecrstFocusReporting => write(self, b"\x1b[?1004l"),
             DecsetBracketedPaste => write(self, b"\x1b[?2004h"),