+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Notable improvements and fixes
 ------------------------------
 - New Taiwanese Chinese translation.
 - :ref:`Transient prompt <transient-prompt>` containing more lines than the final prompt will now be cleared properly (:issue:`11875`).
+- It's now possible to put the cursor by clicking the desired position with your mouse.
 
 Deprecations and removed features
 ---------------------------------